Fix chapter 2 of demo to use correct components

+++ core/util.tsx
@@ -1,6 +1,7 @@
 import { MouseEvent } from 'react'
 import { Toc, TocItem } from 'core/types'
 import * as localForage from "localforage"
+import { Persistor } from 'redux-persist'
 
 /* A function that returns the most-significant word from a phrase,
 typically in the inventory list (e.g. "a tired-looking cap"). Nothing
@@ -27,15 +28,17 @@ export const wordFromInventory = (inventory: string, offset = -1): string => {
 
 
 /* Reset the game and remove the local storage */
-export const resetGame = (userInitiated: boolean, persistor): void => {
+export const resetGame = (userInitiated: boolean, persistor: Persistor): void => {
 
     if (userInitiated) {
         const message = 'Restart the story from the beginning?'
-        persistor.flush().then(() => {
-            persistor.pause()
-            localStorage.clear()
-            window.location.replace(window.location.pathname)
-        })
+        if (confirm(message)) {
+            persistor.flush().then(() => {
+                persistor.pause()
+                localStorage.clear()
+                window.location.replace(window.location.pathname)
+            })
+        }
     }
     else {
         persistor.flush().then(() => {
@@ -49,4 +52,3 @@ export const resetGame = (userInitiated: boolean, persistor): void => {
 export const getChapter = (toc: Toc, filename: string): TocItem => (
     Object.values(toc).filter(c => c.filename === filename)[0]
 )
-
